fix wrong proto and protobufs (#237)
This commit fixes the following issues:

1. In proto, TransactionStatus has been changed from enum
to message and it includes a map of string to tx status.
However, the corresponding protobuf message is not updated
as it already contained the correct TransactionStatus struct.
That's why CI and unit-test passed.

2. In generated grpc protobuf, there is an additional api
called StartCommitInvalidTransactionIDsStream() which is not
used or implemented. Though it is not present in the proto
file, it is present in the committed protobuf. This extra
api and related message, i.e., InvalidTransactionIDs have
been removed from the generated protobufs. As these are dead
code, no failure was observed in CI and unit-test.

Signed-off-by: senthil <snatara7@in.ibm.com>