Include txid in transaction proto (#272)
* signature verifier management

This commit add a signature verifier manager at the coordinator
service. The task of this manager is to manage the communication line
between the coordinator and signature verifier. Using the communication
line, the manager would send txs for signature verification and
receive the result.

Signed-off-by: senthil <snatara7@in.ibm.com>

* include txid in transaction

As part of the transaction, the client needs to include a txID.
This commit introduces txID into the transaction proto message.

Signed-off-by: senthil <snatara7@in.ibm.com>