reduce #keys in dependency detector (#334)
This commit reduces the total number of keys
kept in the dependency detector to detect
dependencies. This is done by introducing
a readWrite key map along with readOnly key map and
writeOnly key map. This readWrite key map hold keys being
both read and written. Earlier, the same key
was added to read key map and write key map but
with this commit, the key will be added only to
read-write key map once. As a result, with 2 inputs and
2 outputs, the throughput of coordinator increased from
170k tps to 270 tps that too at 500 ms latency.

Signed-off-by: senthil <snatara7@in.ibm.com>