retry when error returned from db (#335)
YugabyteDB returns certain errors which are called retryable
error. This means that the client needs to resubmit the transaction.
Though this occur may occur very rarely, we need to handle them.
This commit make the committer to retry three times whenever an error
is returned by the db. Note that we do not classify errors as retryable or not
but simply retry 3 times for every error.

Current set of retryable errors are listed at
https://support.yugabyte.com/hc/en-us/articles/4409627048461-How-to-Troubleshoot-Database-Transaction-Retryable-Errors

Signed-off-by: senthil <snatara7@in.ibm.com>