fix yuga command (#405)
This commit removes the --advertise-address 0.0.0.0 from the
yugabyte start command. The podman container maps 127.0.0.1
from the host to the container, and hence, the host cannot connect
to the Yugabyte server. With old version of yugabyte, i.e., with
yugabyte:2.20.0.1-b1, this is not an issue. However, with the
latest version of yugabyte, this results in a connection failure.
Considering that the test setup is to be used with local Docker,
we can remove the --advertise-address 0.0.0.0 from the start command.

Signed-off-by: senthil <snatara7@in.ibm.com>