fix data race in dependency graph (#413)
This commit fixes a data race in the dependency graph package.
The dependency detector had a race condition when we executed
`ensureEmptyDetector()` in the test while the dependency detector
was being emptied. While this can be addressed using an empty
`atomic.Bool`, we follow a simple approach as this occurs only
in the test. We fix the data race by using metrics to ensure that
all transactions are processed before calling `ensureEmptyDetector()`.
As a result, we ensure that the dependency detector is emptied
before making this call.

Signed-off-by: senthil <snatara7@in.ibm.com>