rename functions in channel utils (#429)
We have functions such as NewInputChannel, NewOutputChannel
but we are creating a reader and writer for the channel
being passed. This is very similar to the Reader and Writer
interface in io package which can take bytes or buffer as
a parameter. Hence, this commit renames NewInputChannel,
CewOutputChannel to channel.NewReader and channel.NewWriter.

Signed-off-by: senthil <snatara7@in.ibm.com>