vcservice: check for pending txs for processing (#446)
* vcservice: check for pending txs for processing

This commit introduces a grpc API to check whether a vcservice
has completed processing all requests. The processing refers to
whether the tx has been fully validated/committed-or-aborted
along with the status. This API would be used by the coordinator
when it is recovering from a failure. Only when all vcservice
are idle, the coordinator can start the recovery process.

Signed-off-by: senthil <snatara7@in.ibm.com>

* address review comments

Signed-off-by: senthil <snatara7@in.ibm.com>

---------

Signed-off-by: senthil <snatara7@in.ibm.com>