sidecar: make errgroup goroutine to return err (#487)
Only when the goroutine in the errgroup returns an error, the context is
cancelled. This commit makes the goroutines which use the stream to
return err so that other goroutines which relay purely on the context
can return too. When the goroutine in
errgroup returns due to context cancellation, it would return a nil
error.

Signed-off-by: senthil <snatara7@in.ibm.com>