coordinator: wait for connections (#526)
#### Type of change

- Bug fix

 
#### Description

This commit makes the coordinator cmd to wait for connections to be
established with sigverifiers and vcservices.

#### Additional details (Optional)

During end-to-end deployment with endorserX, we observed a sidecar
failure while setting the last committed block number. This occurred
because connections to vcservices were not ready in the coordinator
causing the gRPC call to fail and subsequently halt the sidecar. While a
retry mechanism in the sidecar could prevent this specific failure, a
more robust solution is needed to address similar issues proactively.
Hence, this PR ensures that grpc service at the coordinator is started
only after all connections are ready.

We yet to handle failure scenarios where all vcservices or sigverifiers
fail and no connection is accessible.

#### Related issues

Resolves #525

Signed-off-by: senthil <snatara7@in.ibm.com>