add blk#,tx# in the tx status table (#524)
#### Type of change

- New feature

Failure/Recovery
 
#### Description

This commit adds blk# and tx# to each status in the tx_status table. 

#### Additional details (Optional)

This will be used during coordinator recovery. The plan is to remove the
last seen block number and resubmit all transactions from the last
committed block number plus one onward. When a duplicate txID is found,
we will fetch the blk# and tx# to determine if the entry is for a past
commit of the same transaction or if it is truly a duplicate
transaction. To implement this recovery logic, we need to include blk#
and tx# in the status table for each entry.


#### Related issues

Resolves #482

Signed-off-by: senthil <snatara7@in.ibm.com>