coordinator: split large block into chunks (#536)
#### Type of change

- Improvement (performance)
 
#### Description

"During performance benchmarking, it was discovered that a block size of
500 provides optimal performance compared to larger block sizes,
specifically for the dependency graph component. Given that the ordering
service can utilize a larger block size, we need to split it into chunks
within the coordinator to ensure dependency graph performance is not
impacted. This commit splits large blocks into chunks of 500
transactions."

#### Additional details (Optional)

When moving from block size 500 to 5000, we see 10% drop in throughput.
With this PR, we can avoid this degradation.

#### Related issues

Resolves #535

Signed-off-by: senthil <snatara7@in.ibm.com>