make sidecar query next expected block number from coordinator (#550)
#### Type of change

- New feature
 
#### Description

When the sidecar reconnects to the coordinator after a coordinator
failure or a sidecar failure, it should not fetch from block 0. Instead,
it must contact the coordinator to determine the next expected block
number. The sidecar should then fetch blocks from the ordering service
starting at that number.

#### Additional details (Optional)

We are not checking the last committed block number in the block store
managed by the sidecar. This is because we plan to remove the block
store to make the sidecar stateless as part of issue #549. As a result,
the block store could fall behind the coordinator's last committed block
number. This behavior is acceptable because the recovery feature is not
yet complete.

#### Related issues

Resolves #504

---------

Signed-off-by: senthil <snatara7@in.ibm.com>