Resolve vcservice's resource leak, inconsistent and unsafe behaviour during initialization (#559)
**The PR addresses the following changes:**

#### Type of change

- Bug fix
- Improvement (improvement to code, performance, etc)
- Test update
 
#### Description

* Creating a vc-service struct without calling its Run() method will
cause a resource leak. Therefore, the changes in this PR move all
runnable resources into the Run() method, ensuring they are properly
closed at the end.

* The vc-service may respond to gRPC calls before it is fully
initialized (this fix is intended to prevent issues that could arise
from the above changes). The problem has been resolved as described in
#552

#### Related issues

This PR solves issue #546 and partially #553