Solve various issues that cause flaky tests (#561)
#### Type of change

- Bug fix
- Refactoring/Code-Improvement
 
#### Description

This PR addresses a few issues that cause flaky tests.
1. Tests end without ensuring all background activities are stopped.
This was addressed well in most places but was not always handled well
in the `loadgen`, mocks, and integration tests.
This PR adds common test methods to handle the test's background
activities and refactor the code to use it in all tests I found that use
background activities.
This will allow us to easily correct such issues in a single place in
the code.
2. Multiple Go routines used a single load generator. The workload
generators do not support parallelism. Instead, we make a new generator
for each routine that consumes the same stream.
3. The `loadgen` client ended without stopping background activities and
closing connections.
This issue was fixed by applying the service API to the `loadgen`
package, allowing safe resource cleanup.
4. Unnecessary closing of stream
([source](https://github.com/grpc/grpc-go/commit/365770fcbd7dfb9d921cb44827ede770f33be44f)).
Sometimes, calling `stream.CloseSend()` causes halts. Since it is
sufficient to close the connection, these calls are removed.
5. The `sidecar` client did not properly verify that the namespace TX
was committed.
This caused the `loadgen` client and integration tests to fail.
This was solved by counting the difference of committed TXs from before
submitting the namespace TX.
Also, we omit config TXs from the TX count as they are not generated by
the loadgen.
6. Add `Close()` method to the ledger-service and the sidecar to allow
releasing the ledger directory.

Since we already made significant modifications to the `loadgen` and the
`mock`, this PR also refactored them.

This PR refactors the mocks to reside in the `mock` package.

This PR refactors the `loadgen` package to use 3 sub-packages: 
1. `workload`: generates the transactions and the blocks
2. `adapters`: applies a load on a specific service
3. `metrics`: performance metrics
4. `loadgen` (main): the client that integrates the `workload` and
`adapters`

This PR also removes the unnecessary code duplication for creating the
namespace and applying the main workload.
This allows instantiating the service adapter (and all its connections)
once for the entire run, instead of twice (once for the namespace and
once for the workload).

#### Additional details

Apologies for the big PR in terms of the number of files.
The reviewer should note that most of the changes are syntactic.
However, the `loadgen`, `integration`, and `mock` packages had some
semantic changes to correct the behavior that caused flaky tests.

Some flaky tests may still occur due to the use of the `loadgen` package
in the query service tests. This will be further investigated in the
future.

#### Related issues

Refactor:
resolve #496 
resolve #501 

Flaky Tests:
resolve #396 
resolve #519 
resolve #400
resolve #473
resolve #477

---------

Signed-off-by: Liran Funaro <liran.funaro@gmail.com>
Co-authored-by: Liran Funaro <liran.funaro@gmail.com>