Make GetTransactionsStatus return status with height info (#637)
#### Type of change

- New feature
- Improvement to the code. 
 
#### Description

This commit adds GetTransactionsStatus gRPC api in the coordinator. In
addition to status, we make this api return the block number and
transaction number too.

Further, this commit replaces
`protocoordinatorservier.TransactionsStatus`, and
`protovcservice.TransactionStatus` with a single
`protoblocktx.TransactionStatus`.


#### Additional details (Optional)

This is required to recover sidecar ledger store. This API will be used
in the sub-sequent PR that recovers sidecar ledger store.

#### Related issues

Resolves #528

---------

Signed-off-by: senthil <snatara7@in.ibm.com>