Fix Sidecar recovery test and add ctx to the RunServiceForTest utility. (#640)
#### Type of change

- Bug fix
 
#### Description

This commit addresses an issue with the recovery test in TestSidecar.
Previously, the test used a custom context to stop the service
completely. However, when RunServiceForTest() was introduced, the custom
context was removed, causing the test to only stop the gRPC service
rather than the entire service. This behavior was insufficient for the
recovery test.

To resolve this, the commit reverts the relevant changes in TestSidecar
and updates RunServiceForTest() to accept a context. This allows us to
perform comprehensive recovery tests for all components.

#### Related issues

Resolves #639

---------

Signed-off-by: senthil <snatara7@in.ibm.com>