[vcservice] allow tx resubmissions and fix a flaky test  (waiting for ready in the coordinator) (#650)
#### Type of change

- Bug fix
- New feature
 
#### Description

To simplify the coordinator’s recovery after a vcservice failure, this
commit enable the vcservice to handle transaction resubmissions.

Consider the following scenario: a connection from the coordinator to
the vcservice may fail due to either

1. a network partition or 
2. a server crash. 

If the server crashes, there is no issue. However, if it is a network
partition, the coordinator might resubmit the same transactions to a
different vcservice instance while the disconnected one is still
processing those transactions. In a corner case, once reconnected, the
coordinator might submit the same transactions yet again to the
previously disconnected vcservice.

By making the vcservice return a consistent result for all transaction
resubmissions, we can simplify the coordinator’s recovery logic. Thus,
the coordinator can simply requeue the pending transactions to the input
queue so that any available vcservice can process them.

#### Additional details (Optional)

This commit also fix a bug in the coordinator. For details, refer to
issue #663

#### Related issues

Resolves #663
Partially address #502

---------

Signed-off-by: senthil <snatara7@in.ibm.com>