[coordinator] skip verifier that has connection issue and update policies during connection re-establishment  (#695)
#### Type of change

- New feature
 
#### Description

If a verifier connection is temporarily unavailable, the
`updatePolicies()` call would previously fail, causing the coordinator
to crash. This commit changes that behavior. Now, the signature verifier
manager updates policies in the background when a verifier is
unreachable, preventing coordinator crashes due to transient connection
issues. The coordinator will no longer crash even if all verifiers are
temporarily unavailable. A crash will only occur if a verifier returns
an internal error indicating a serious problem.

Upon connection re-establishment, policies are updated before any
requests are sent to the verifier.

#### Additional details (Optional)

1. Wraps are added to fix a few lints.
2. A few other fixes to make linter happy. 

#### Related issues

Resolves #582
Resolves #665

---------

Signed-off-by: senthil <snatara7@in.ibm.com>