[coordinator] decouple txBatchID used for dependency graph from the stream (#759)
#### Type of change

- Bug fix
 
#### Description

We do send a `txBatchID` which can be different from the block number to
the dependency graph so that the graph manager can order the received
transactions correctly. So far, we have coupled the `txBatchID` with the
stream and always set it to 1. This does not work when the sidecar
reconnects to coordinator due to a connection issue or sidecar restart.
Hence, this commit moves the `txBatchID` to the coordinator as a field
in the coordinator struct.

#### Related issues

Resolves #757

Signed-off-by: senthil <snatara7@in.ibm.com>