[VC] set retry policy from DBConfig to DB object (#850)
#### Type of change

- Bug fix

#### Description

The database configuration includes a retry policy, but it is never
applied to the database object. As a result, the database object always
uses the default reset policy. This PR assigns the user defined retry
policy to the DB object.

#### Additional details (Optional)

This PR also adds the retry configuration to the sample file but not to
the template (as it is not needed there).

#### Related issues

 - Resolves #685

Signed-off-by: Senthil Nathan N <snatara7@in.ibm.com>