[Coordinator] Make VCM not forward status of untracked txIDs (#846)
#### Type of change

- Bug fix
 
#### Description

We send statuses to coordinator immediately on receiving it from the vc
service. However, we should only forward them if the respective
transactions are part of txBeingValidated. Otherwise, we might forward
duplicate statuses during failure/recovery. This PR checks whether the
status received for a transaction is being tracked. If not, the status
is ignored.

#### Related issues

 - Resolves #843

---------

Signed-off-by: Senthil Nathan N <snatara7@in.ibm.com>