[runner] Start loadgen within the Start() itself (#848)
#### Type of change

- Improvement (improvement to code)
 
#### Description

1. For `LoadGenForCommitter()`, we need to call
`StartLoadGenForCommitter()` before the `StartSystem()`. Otherwise, the
test does not terminate though it passes successfully.
2. For `LoadGenForOrderer()`, the invocation order does not matter.

This is non-intuitive and might require some time to understand the flow
for the first time. As loadGen is also part of the `SystemConfig`, we
start it within the `Start()` itself with necessary flags.

#### Additional details (Optional)

Further, the query service is not required for all integration test.
Hence, we remove this flag for most test.

#### Related issues

 - Resolves #847 
 - Resolves #849

---------

Signed-off-by: Senthil Nathan N <snatara7@in.ibm.com>