[BREAKING] Fix TX ASN1 digest (#64)
#### Type of change

- Bug fix
- Test update
- Breaking change
 
#### Description

- Add namespace ID and version to the ASN.1 schema
- Remove redundant marshaling when using NONE schema
  - Allows better benchmarking of non-sig related components
- Add a test that validates the marshalling schema includes all the
fields of the TX by reconstructing the TX from the marshalled data
- Add fuzz test for the ASN.1 marshalling
- Add a benchmark for the digest method

The tests discovered the following:
1. The ASN.1 library may encode a string in two ways, depending on the
input string.
If it is ASCII, it encodes it as PrintableString; otherwise, it uses
UTF-8.
This may cause incompatibility with the schema, which requires a
specific type. Thus, we force the type to be UTF-8.
2. The ASN.1 marshalling cannot differentiate between `nil` and empty
slice (`[]byte{}`).
We must ensure our implementation does not make this differentiation to
avoid inconsistencies.
The current implementation parses version `[]byte{}` as zero. If the
endorser signed on `nil` version, an attacker can change the TX to have
version zero by replacing the `nil` with `[]byte{}` and go unnoticed.

To fix these issues, we:
- Modify the ASN.1 schema to use UTF-8 (UTF8String) instead of ASCII
(PrintableString)
- Change version to be uint64. 
- Since ASN.1 doesn't support unsigned int, we use int64 and encode nil
as `-1`.
- Since SQL doesn't support unsigned number, we use bigint (int64) and
add check for NULL and version>=0.

#### Related issues

- resolves #53  
- address #52

---------

Signed-off-by: Liran Funaro <liran.funaro@gmail.com>