[coordinator] Productization of the simple dependency manager (#65)
#### Type of change

- Improvement

#### Description

Simple manager:
- Add support for waiting TX limit
- Add metrics 
- Add test
- Optimize `constructCompositeKey()`

#### Performance

Benchmarking was done on an Apple M1 machine.
When dependency was introduced, it was with 30% dependent TXs.
The following tests are conducted with 2, 4, or 6 independent read-write
keys, and dependent keys are added accordingly.
All combinations of dependencies were tested. However, all conflicts
with write had the same performance as write-write conflicts.


![image](https://github.com/user-attachments/assets/9c005872-1779-4b5b-b600-b075e89afe30)

Benchmark on i9-14900K

![image](https://github.com/user-attachments/assets/a74a4eac-f7c8-4297-b637-fbef9e04b24c)

#### Related issues

- resolves #56

---------

Signed-off-by: Liran Funaro <liran.funaro@gmail.com>