fixed dashboards for `loadgen` throughput and committer transactions (#119)
#### Type of change

- Bug fix
 
#### Description

<!--- Describe your changes in detail, including motivation. -->

The `loadgen` Throughput dashboard was not showing the throughput
correctly. The issue was related to the sampling window set on grafana,
e.g. 5sec, which was too little to collect enough samples and allow the
`rate()` function to work properly.
This dashboard has been tested with `fabric-x-committer:0.1.5` and it
works.

Signed-off-by: pco <pasquale.convertini@ibm.com>